refactor(gateway): Q11.b — extract rest_v1/tasks/comments.rs (GAR-635)
Extracts 3 comment handlers + 5 DTOs (~414 LOC) from tasks/mod.rs into
tasks/comments.rs (plan 0136, GAR-635 slice 2). mod.rs drops from 3586
to 3180 lines — below the 3240-line quality-ratchet baseline. Zero
behaviour change; re-exports in mod.rs keep all call-sites intact.
