+++ crates/garraia-gateway/src/rest_v1/openapi.rs
@@ -126,9 +126,9 @@ impl Modify for SecurityAddon {
         super::tasks::get_task,
         super::tasks::patch_task,
         super::tasks::delete_task,
-        super::tasks::create_task_comment,
-        super::tasks::list_task_comments,
-        super::tasks::delete_task_comment,
+        super::tasks::comments::create_task_comment,
+        super::tasks::comments::list_task_comments,
+        super::tasks::comments::delete_task_comment,
         super::tasks::add_task_assignee,
         super::tasks::list_task_assignees,
         super::tasks::remove_task_assignee,
